requirements.txt
selenium == 4.8.3
openai-whisper == 20230314